+++ dashboard/src/app/[locale]/(public)/(landing)/landingPage.tsx
@@ -10,7 +10,6 @@ import { PricingSection } from './components/pricingSection';
 import { OpenSourceCallout } from './components/openSourceCallout';
 import { buildSEOConfig, SEO_CONFIGS } from '@/lib/seo';
 import { CtaStrip } from '@/components/public/ctaStrip';
-import { ScrollStateProvider } from '@/contexts/ScrollStateProvider';
 
 export default async function LandingPage() {
   const seoConfig = await buildSEOConfig(SEO_CONFIGS.landing);
@@ -82,8 +81,7 @@ export default async function LandingPage() {
   return (
     <>
       <StructuredData config={seoConfig} />
-      <ScrollStateProvider>
-        <div className='bg-background text-foreground relative overflow-x-clip'>
+      <div className='bg-background text-foreground relative overflow-x-clip'>
         <div className='flex flex-col'>
           {landingSections.map((section) => {
             const { id, orientation, content, showGradient = true } = section;
@@ -96,8 +94,7 @@ export default async function LandingPage() {
             );
           })}
         </div>
-        </div>
-      </ScrollStateProvider>
+      </div>
     </>
   );
 }
